Rename controller files
