+++ server/src/controllers/index.ts
@@ -1,3 +1,3 @@
-export { TrainingController } from "./base.js";
-export { FederatedController } from "./federated.js";
-export { DecentralizedController } from "./decentralized.js";
+export { TrainingController } from "./training_controller.js";
+export { FederatedController } from "./federated_controller.js";
+export { DecentralizedController } from "./decentralized_controller.js";
