+++ server/src/controllers/training_controller.ts
